Moved demo charts to separate file
